moved fragmentation and library spectrum to new project Omics
