+++ mzLib/Omics/Class1.cs
@@ -1,0 +1,7 @@
+﻿namespace Omics
+{
+    public class Class1
+    {
+
+    }
+}
