+++ mzLib/Proteomics/ProteolyticDigestion/ProteaseDictionary.cs
@@ -4,7 +4,6 @@ using System.Globalization;
 using System.IO;
 using System.Linq;
 using MzLibUtil;
-using Proteomics.Fragmentation;
 
 namespace Proteomics.ProteolyticDigestion
 {
