+++ mzLib/Proteomics/ProteolyticDigestion/ProteinDigestion.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using Proteomics.Fragmentation;
+using Omics.Fragmentation;
 
 namespace Proteomics.ProteolyticDigestion
 {
