+++ mzLib/Test/TestPsmFromTsv.cs
@@ -1,5 +1,4 @@
 ﻿using NUnit.Framework;
-using Proteomics.Fragmentation;
 using Proteomics.ProteolyticDigestion;
 using Proteomics.PSM;
 using System;
@@ -7,6 +6,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text.RegularExpressions;
+using Omics.Fragmentation;
 using Proteomics;
 
 namespace Test
